# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SML Staff Marketplace API
  version: "1.1.0"
  description: |
    Express.js port of the original Java/JAX-RS MarketPlaceWebService.

    All endpoints are under `/service/v1` (included in the server URL below).

    **Auth note**: Failed login returns HTTP 200 with `data: []` — not HTTP 401.

    **Image endpoints**: Return binary `image/png`. `/imagesguid` never returns 404 — returns empty buffer instead.
servers:
  - url: http://localhost:47309/service/v1
    description: Local development (current SML Staff Ubon)
  - url: http://localhost:47300/service/v1
    description: Local development (legacy preview)
...
security:
  - GUIDHeader: []
components:
  securitySchemes:
    GUIDHeader:
      type: apiKey
      in: header
      name: GUID
      description: Session identifier required by most endpoints
  schemas:
...
    ErrorResponseLower:
      type: object
      properties:
        error:
          type: string
...
    BarcodeItem:
      type: object
      properties:
        item_code: { type: string, example: "01-0004" }
        item_name: { type: string, example: "ซีรีแล็ค โจ๊ก 200กรัม/ผัก 3 สี" }
        barcode: { type: string, example: "9556001131751" }
        unit_code: { type: string, example: "ชิ้น" }
        price: { type: number, example: 100 }
        stand_value: { type: number, example: 1 }
        divide_value: { type: number, example: 1 }
        ratio: { type: number, example: 1 }
        stock_qty: { type: number, example: 0 }
        tax_type: { type: integer, example: 0 }
...
paths:
...
  /getBarcodeItem:
    get:
      tags: [Product]
      summary: ค้นหาสินค้าจากบาร์โค้ด
      description: Used by sale and PU manual barcode flows. Returns an empty data array when barcode is not found.
      parameters:
        - { name: barcode, in: query, required: true, schema: { type: string }, example: "9556001131751" }
        - { name: custcode, in: query, schema: { type: string }, description: "Reserved for price/customer-specific logic" }
      responses:
        '200':
          description: Product detail rows for barcode
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/BarcodeItem' }
        '400':
          description: barcode is required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }
        '500':
          description: DB error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }
  /getBarcodeItemSearch:
    get:
      tags: [Product]
      summary: ค้นหาสินค้าจากรหัส ชื่อ หรือบาร์โค้ด
      description: Search dialog endpoint used by sale and PU; empty search returns an empty data array.
      parameters:
        - { name: search, in: query, schema: { type: string }, example: "01" }
        - { name: custcode, in: query, schema: { type: string }, description: "Reserved for price/customer-specific logic" }
        - { name: limit, in: query, schema: { type: integer, default: 30, maximum: 100 } }
      responses:
        '200':
          description: Matching products
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/BarcodeItem' }
        '500':
          description: DB error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }

  # ─── CART ────────────────────────────────────────────────────────────────────
...